+++ src/vs/code/electron-sandbox/browser/issueReporterModel.ts
